// api/health-api.ts
/**
 * Cobo Agentic Wallet Service
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import type { AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios';
import globalAxios from 'axios';
import { Configuration, getServerUrl } from '../configuration';
import { assertParamExists } from '../common';

export class HealthApi {
    private basePath: string;
    private axios: AxiosInstance;
    private configuration?: Configuration;

    constructor(configuration?: Configuration, basePath?: string, axios?: AxiosInstance) {
        this.configuration = configuration;
        this.basePath = basePath ?? configuration?.basePath ?? getServerUrl(configuration?.serverIndex ?? 0, configuration?.serverVariables);
        this.axios = axios ?? globalAxios;
    }

    /**
     * Get default request headers including auth
     */
    private getHeaders(): Record<string, string> {
        const headers: Record<string, string> = {};
        if (this.configuration?.apiKey) {
            headers['X-API-Key'] = this.configuration.apiKey;
        }
        return headers;
    }

    /**
     * Health Check
     * @param options Override http request option.
     */
    public async healthCheck(options?: AxiosRequestConfig): Promise<AxiosResponse<{ [key: string]: string; }>> {
        let localVarPath = '/health';
        const localVarUrlObj = new URL(localVarPath, this.basePath);
        const localVarRequestOptions: AxiosRequestConfig = { method: 'GET', ...options };
        const localVarHeaderParameter: Record<string, string> = { ...this.getHeaders() };
        const localVarQueryParameter: Record<string, any> = {};

        localVarHeaderParameter['Content-Type'] = 'application/json';
        localVarHeaderParameter['Accept'] = 'application/json';

        const queryString = new URLSearchParams(
            Object.entries(localVarQueryParameter).reduce((acc, [key, value]) => {
                if (Array.isArray(value)) {
                    value.forEach(v => acc.push([key, String(v)]));
                } else {
                    acc.push([key, String(value)]);
                }
                return acc;
            }, [] as [string, string][])
        ).toString();
        if (queryString) {
            localVarUrlObj.search = queryString;
        }

        localVarRequestOptions.headers = { ...localVarHeaderParameter, ...options?.headers };
        localVarRequestOptions.url = localVarUrlObj.toString();

        return this.axios.request(localVarRequestOptions);
    }
}