// models/pending-operation-status.ts
/**
 * Cobo Agentic Wallet Service
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */






/**
 * Approval workflow status.
 * @export
 * @enum {string}
 */
export const PendingOperationStatus = {
    pending: 'pending' as const,
    executing: 'executing' as const,
    rejected: 'rejected' as const,
    executed: 'executed' as const,
} as const;

export type PendingOperationStatus = typeof PendingOperationStatus[keyof typeof PendingOperationStatus];